*Fixed DO-1128 Complete Registration and reset password email does not g...

+++ api/src/main/java/org/ednovo/gooru/domain/service/user/impl/UserServiceImpl.java
@@ -1105,7 +1105,7 @@ public class UserServiceImpl implements UserService,ParameterProperties,Constant
 		Identity identity = this.getUserRepository().findIdentityByResetToken(token);
 		boolean resetTokenInvalid = false;
 		if (identity != null) {
-			long resetHoursDifferent = (new Date(System.currentTimeMillis()).getTime() - identity.getCredential().getResetPasswordRequestDate().getTime()) / (60 * 60 * 1000);
+			double resetHoursDifferent = (new Date(System.currentTimeMillis()).getTime() - identity.getCredential().getResetPasswordRequestDate().getTime()) / (60 * 60 * 1000);
 			if (resetHoursDifferent > 24) {
 				resetTokenInvalid = true;
 			}
